+++ ApplicationCode/ReservoirDataModel/RigSimulationWellCenterLineCalculator.h
@@ -33,7 +33,7 @@ public:
                                                   std::vector< std::vector <RigWellResultPoint> >& pipeBranchesCellIds) ;
 
     static void calculateWellPipeDynamicCenterline(const RimSimWellInView* rimWell, 
-                                                   size_t timeStepIndex,
+                                                   int timeStepIndex,
                                                    std::vector< std::vector <cvf::Vec3d> >& pipeBranchesCLCoords,
                                                    std::vector< std::vector <RigWellResultPoint> >& pipeBranchesCellIds) ;
